refactor: Update API generation scripts to use new endpoints, as they couldn't handle redirects. consolidate API generation scripts to TypeScript only
